helper/old _server/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
celery
redis
django-redis